requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
beautifulsoup4>=4.9.3
pinecone-client>=2.2.1
openai>=1.0.0
python-dotenv>=1.0.0 